Try to create a working version

+++ bin/chef-foodcritic-publisher.sh
@@ -1,11 +1,24 @@
 #!/bin/bash
-mkdir -p junit_reports
+
+if [ ! -d cookbooks ]; then
+    echo 'This script must be run from the root of the chef repository'
+    exit 1
+fi
+
+if [ ! -d junit_reports ]; then
+  mkdir -p junit_reports
+fi
+
 rm junit_reports/foodcritic-*.xml 2>/dev/null
 
-source /usr/local/rvm/scripts/rvm
-rvm use ruby-1.9.3 2>/dev/null
+PATH=${HOME}/bin:${PATH}
+
+if [ -s ${HOME}/.rvm/scripts/rvm ]
+then
+    . "$HOME/.rvm/scripts/rvm" # Load RVM into a shell session *as a function*
+fi
 
 for cbname in `find cookbooks -maxdepth 1 -mindepth 1 -type d | sed -e 's/cookbooks\///'`; do
   echo "------ foodcritic checks: $cbname ------"
-  foodcritic --tags ~FC001 --tags ~FC020 --tags ~FC035 $@ cookbooks/$cbname | foodcritic2junit.pl --suite $cbname --out junit_reports/foodcritic-$cbname.xml
+  foodcritic --tags '~FC011' --tags '~FC015' --tags="~FC033" --tags="~FC034" --tags "~FC005" $@ cookbooks/$cbname | chef-ci-tools/bin/foodcritic2junit.pl --suite $cbname --out junit_reports/foodcritic-$cbname.xml
 done
